apps/api-crawler/requirements.txt
faststream[redis]
faststream[cli]
pymongo
dotenv
APScheduler
-e file:../../packages/shared
-e file:../../packages/gw2-api